+++ src/Components/CreateImageWizard/steps/Oscap/components/useSelectorHandlers.tsx
@@ -1,12 +1,19 @@
+import { v4 as uuidv4 } from 'uuid';
+
 import { useAppDispatch } from '../../../../../store/hooks';
-import { Services } from '../../../../../store/imageBuilderApi';
+import { Filesystem, Services } from '../../../../../store/imageBuilderApi';
 import {
   addPackage,
+  addPartition,
   changeDisabledServices,
   changeEnabledServices,
+  changeFileSystemConfigurationType,
   changeMaskedServices,
+  clearPartitions,
   removePackage,
 } from '../../../../../store/wizardSlice';
+import { parseSizeUnit } from '../../../utilities/parseSizeUnit';
+import { Partition, Units } from '../../FileSystem/components/FileSystemTable';
 
 export const useSelectorHandlers = () => {
   const dispatch = useAppDispatch();
@@ -35,6 +42,28 @@ export const useSelectorHandlers = () => {
     }
   };
 
+  const handlePartitions = (oscapPartitions: Filesystem[]) => {
+    dispatch(clearPartitions());
+
+    const newPartitions = oscapPartitions.map((filesystem) => {
+      const [size, unit] = parseSizeUnit(filesystem.min_size);
+      const partition: Partition = {
+        mountpoint: filesystem.mountpoint,
+        min_size: size.toString(),
+        unit: unit as Units,
+        id: uuidv4(),
+      };
+      return partition;
+    });
+
+    if (newPartitions.length > 0) {
+      dispatch(changeFileSystemConfigurationType('manual'));
+      for (const partition of newPartitions) {
+        dispatch(addPartition(partition));
+      }
+    }
+  };
+
   const handleServices = (services: Services | undefined) => {
     dispatch(changeEnabledServices(services?.enabled || []));
     dispatch(changeMaskedServices(services?.masked || []));
@@ -45,5 +74,6 @@ export const useSelectorHandlers = () => {
     clearCompliancePackages,
     handlePackages,
     handleServices,
+    handlePartitions,
   };
 };
